+++ Project1_Character_Library/src/Controller/GeneralViewerController.java
@@ -68,7 +68,7 @@ public class GeneralViewerController {
     }
     
     public Character createCharacter(String pName, int pLife, int pLevelReq,int pLevel, double pHitsPerTime, double pFields,
-                                Direction pDirection,int pCost) throws IOException {
+                                Direction pDirection,double pCost) throws IOException {
         return characters.createCharacter(pName, pLife, pLevelReq, pLevel, pHitsPerTime, pFields, pDirection,pCost);
     }
